+++ index.html
@@ -3,7 +3,7 @@
   <head>
     <script
       language="javascript"
-      src="index.js"
+      src="./dist/bundle.js"
       type="text/javascript"
     ></script>
     <title>js-project</title>
