+++ src/connection.js
@@ -1,0 +1,5 @@
+const GITHUB_API_URL = 'https://api.github.com';
+function getGithubRepoSearchUrl(query) {
+  return `${GITHUB_API_URL}/search/repositories?q=${query}&page=1&per_page=10`;
+}
+  export {getGithubRepoSearchUrl};
