+++ webpack.config.js
@@ -1,0 +1,16 @@
+const path = require('path');
+
+module.exports = {
+    entry:[ 
+        "./src/index.js",
+        "./src/connection.js",
+        "./src/domHandler.js"
+    ]
+
+    ,
+    output:{
+        path:path.resolve(__dirname,"./dist"),
+        filename: "bundle.js"
+    },
+    mode: "production"
+}
